Ensure evaluation on pinhole datasets (classic, hopev2) works without hand_tracking_toolkit installation
